+++ src/components/Header/SessionControl/style.scss
@@ -7,6 +7,7 @@
         background-color: rgba($-dark, 0.1);
         border: none;
         box-shadow: none;
+        color: $-light;
 
         &::after {
             vertical-align: 0.1em;
