+++ src/dynamic/Logs/index.jsx
@@ -1,14 +1,14 @@
 import React from "react";
 import { Container, Badge } from "react-bootstrap";
 
-function Settings() {
+function Logs() {
   return (
     <Container className="py-5">
       <h2>
-        Settings <Badge variant="primary">Coming Soon</Badge>
+        Logs <Badge variant="primary">Coming Soon</Badge>
       </h2>
     </Container>
   );
 }
 
-export default Settings;
+export default Logs;
