+++ src/dynamic/Statistics/index.jsx
@@ -1,14 +1,14 @@
 import React from "react";
 import { Container, Badge } from "react-bootstrap";
 
-function Settings() {
+function Statistics() {
   return (
     <Container className="py-5">
       <h2>
-        Settings <Badge variant="primary">Coming Soon</Badge>
+        Statistics <Badge variant="primary">Coming Soon</Badge>
       </h2>
     </Container>
   );
 }
 
-export default Settings;
+export default Statistics;
