+++ src/pages/Index/style.scss
@@ -180,14 +180,6 @@
             }
         }
 
-        .badge {
-            font-weight: 400;
-            padding: 0.4em 0.6em;
-            border-radius: 4px;
-            margin-left: 0.4rem;
-            margin-top: 0.3em;
-        }
-
         .minor-feature--icon {
             display: block;
             width: 80px;
