+++ src/scss/main.scss
@@ -96,3 +96,11 @@ main > div[role="group"] {
 .modal-backdrop.show {
     opacity: 0.65 !important;
 }
+
+.badge {
+    font-weight: 400;
+    padding: 0.35em 0.6em 0.4em;
+    border-radius: 4px;
+    margin-left: 0.4em;
+    margin-top: 0.2em;
+}
