+++ src/store/api/labels.js
@@ -1,3 +1,4 @@
 export const TOKEN_EXCHANGE = "TOKEN_EXCHANGE";
 export const IDENTIFY_SESSION = "IDENTIFY_SESSION";
 export const GET_GUILDS = "GET_GUILDS";
+export const GET_RESPONSES = "GET_RESPONSES";
