+++ src/store/reducers/guilds.js
@@ -1,12 +1,13 @@
 import { createSelector } from "reselect";
-import { LOAD_GUILDS, SIGN_OUT } from "../actions";
+import { LOAD_GUILDS } from "../actions";
+import { sessionAware } from "./session";
 
 export const initial = {
   guildList: [],
   hasLoaded: false
 };
 
-export function reducer(state = initial, action) {
+export const reducer = sessionAware(initial, (state, action) => {
   switch (action.type) {
     case LOAD_GUILDS:
       const { guildList } = action.payload;
@@ -16,13 +17,10 @@ export function reducer(state = initial, action) {
         hasLoaded: true
       };
 
-    case SIGN_OUT:
-      return initial;
-
     default:
       return state;
   }
-}
+});
 
 const getGuildList = state => state.guilds.guildList;
